requirements.txt
# pip freeze
astroid==2.1.0
attrs==17.4.0
botocore==1.12.82
cachetools==3.0.0
certifi==2018.11.29
chalice==1.6.2
chardet==3.0.4
click==6.7
docutils==0.14
enum-compat==0.0.2
google-api-python-client
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
httplib2==0.12.0
idna==2.8
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pylint==2.2.2
python-dateutil==2.7.5
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.0
